Add modules and classes to JS
